gunicorn/requirements.txt
Django
gunicorn
python-dateutil
psycopg2-binary
django-ratelimit